Move `ConfigurationCacheBuildEnablement` to `o.g.cc.initialization`
